@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <RootNamespace>nexRemote.Desktop.Core</RootNamespace>
     <AssemblyName>nex-Remote_Desktop.Core</AssemblyName>
     <Platforms>AnyCPU;x64;x86</Platforms>
