@@ -1,16 +1,16 @@
 ﻿@attribute [Authorize]
 @inherits AuthComponentBase
 
-<h3>Urządzenia</h3>
+<h3>Devices</h3>
 
 <div @onclick="ClearSelectedCard">
     <div id="deviceControlsWrapper">
         <div>
             <div>
-                <div>Grupa urządzenia</div>
+                <div>Device Group</div>
                 <select class="form-control" style="width: 200px" @bind="_selectedGroupId">
-                    <option value="@_deviceGroupAll">Wszystkie</option>
-                    <option value="@_deviceGroupNone">Żadna</option>
+                    <option value="@_deviceGroupAll">All</option>
+                    <option value="@_deviceGroupNone">None</option>
                     @foreach (var group in _deviceGroups)
                     {
                         <option @key="group.ID" value="@group.ID">@group.Name</option>
@@ -19,7 +19,7 @@
             </div>
 
             <div>
-                <div class="mt-1">Sortuj</div>
+                <div class="mt-1">Sort</div>
                 <select class="form-control d-inline" style="width: 150px" @bind="_selectedSortProperty">
                     <option value=""></option>
                     @foreach (var prop in _sortableProperties)
@@ -27,28 +27,28 @@
                         <option value="@prop.Name">@GetDisplayName(prop)</option>
                     }
                 </select>
-                <button class="btn btn-sm btn-secondary" title="Kierunek Sortowania" style="width: 40px; margin-left: 5px;" @onclick="ToggleSortDirection">
+                <button class="btn btn-sm btn-secondary" title="Sort Direction" style="width: 40px; margin-left: 5px;" @onclick="ToggleSortDirection">
                     <span class="oi @GetSortIcon()"></span>
                 </button>
             </div>
 
             <div class="custom-control custom-checkbox mt-2">
                 <input id="hideOfflineDevicesCheckbox" type="checkbox" class="custom-control-input" @bind="_hideOfflineDevices" />
-                <label class="custom-control-label" for="hideOfflineDevicesCheckbox">Ukryj urządzenia offline</label>
+                <label class="custom-control-label" for="hideOfflineDevicesCheckbox">Hide offline devices</label>
             </div>
             <div class="mt-2">
-                <button class="btn btn-sm btn-secondary mr-2" @onclick="SelectAllCards">Wybierz Wszystko</button>
-                <button class="btn btn-sm btn-primary" @onclick="HandleRefreshClicked">Odśwież</button>
+                <button class="btn btn-sm btn-secondary mr-2" @onclick="SelectAllCards">Select All</button>
+                <button class="btn btn-sm btn-primary" @onclick="HandleRefreshClicked">Refresh</button>
             </div>
         </div>
-         
+
         <div>
             <div class="d-inline-block text-left">
                 <div>
-                    <div>Szukaj</div>
+                    <div>Search</div>
                     <input class="form-control d-inline"
                            style="width: 200px"
-                           placeholder="Szukaj właściwości urządzenia"
+                           placeholder="Search device properties"
                            @bind="_filter" />
                 </div>
 
@@ -56,9 +56,9 @@
                     <button class="btn btn-sm btn-secondary" @onclick="PageDown">
                         <i class="oi oi-chevron-left"></i>
                     </button>
-                    <span class="small">Strona</span>
+                    <span class="small">Page</span>
                     <input id="currentPageInput" style="width:75px; text-align: center" pattern="[0-9]" @bind="_currentPage" @bind:event="oninput" />
-                    <span class="small">z</span>
+                    <span class="small">of</span>
                     <span>@TotalPages</span>
                     <button class="btn btn-sm btn-secondary" @onclick="PageUp">
                         <i class="oi oi-chevron-right"></i>
@@ -72,10 +72,10 @@
                         <option value="100">100</option>
                         <option value="200">200</option>
                     </select>
-                    <span class="ml-1 small">na stronę</span>
+                    <span class="ml-1 small">per page</span>
                 </div>
             </div>
-           
+
 
         </div>
     </div>
