@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using nexRemote.Agent.Interfaces;
+using Moq;
 using nexRemote.Server.Services;
 using nexRemote.Shared.Models;
 using nexRemote.Shared.Utilities;
@@ -15,8 +15,8 @@ namespace nexRemote.Tests
     public class DataServiceTests
     {
         private IDataService _dataService;
-        private IDeviceInformationService _deviceInfo;
         private TestData _testData;
+        private string _newDeviceID = "NewDeviceName";
 
         [TestMethod]
         public async Task AddAlert()
@@ -29,19 +29,25 @@ namespace nexRemote.Tests
         }
 
         [TestMethod]
-        public async Task AddOrUpdateDevice()
+        public void AddOrUpdateDevice()
         {
-            var newDeviceID = "NewDeviceName";
-            var storedDevice = _dataService.GetDevice(newDeviceID);
+            var storedDevice = _dataService.GetDevice(_newDeviceID);
 
             Assert.IsNull(storedDevice);
 
-            var newDevice = await _deviceInfo.CreateDevice(newDeviceID, _testData.OrganizationID);
+            var newDevice = new Device()
+            {
+                ID = _newDeviceID,
+                OrganizationID = _testData.OrganizationID,
+                DeviceName = Environment.MachineName,
+                Is64Bit = Environment.Is64BitOperatingSystem
+            };
+
             Assert.IsTrue(_dataService.AddOrUpdateDevice(newDevice, out _));
 
-            storedDevice = _dataService.GetDevice(newDeviceID);
+            storedDevice = _dataService.GetDevice(_newDeviceID);
 
-            Assert.AreEqual(newDeviceID, storedDevice.ID);
+            Assert.AreEqual(_newDeviceID, storedDevice.ID);
             Assert.AreEqual(Environment.MachineName, storedDevice.DeviceName);
             Assert.AreEqual(Environment.Is64BitOperatingSystem, storedDevice.Is64Bit);
         }
@@ -171,7 +177,14 @@ namespace nexRemote.Tests
         {
             _testData = new TestData();
             _dataService = IoCActivator.ServiceProvider.GetRequiredService<IDataService>();
-            _deviceInfo = IoCActivator.ServiceProvider.GetRequiredService<IDeviceInformationService>();
+
+            var newDevice = new Device()
+            {
+                ID = _newDeviceID,
+                DeviceName = Environment.MachineName,
+                Is64Bit = Environment.Is64BitOperatingSystem,
+                OrganizationID = _testData.OrganizationID
+            };
         }
 
         [TestMethod]
