@@ -3,16 +3,18 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Primitives;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using nexRemote.Agent.Interfaces;
-using nexRemote.Agent.Services;
 using nexRemote.Server.API;
 using nexRemote.Server.Data;
 using nexRemote.Server.Services;
 using nexRemote.Shared.Models;
 using nexRemote.Shared.Utilities;
 using System;
+using System.Collections.Generic;
+using System.Configuration;
 
 namespace nexRemote.Tests
 {
@@ -28,7 +30,14 @@ namespace nexRemote.Tests
             {
                 builder = WebHost.CreateDefaultBuilder()
                    .UseStartup<Startup>()
-                   .CaptureStartupErrors(true);
+                   .CaptureStartupErrors(true)
+                   .ConfigureAppConfiguration(config =>
+                   {
+                       config.AddInMemoryCollection(new Dictionary<string, string>()
+                       {
+                           ["ApplicationOptions:DBProvider"] = "InMemory"
+                       });
+                   });
 
                 builder.Build();
             }
@@ -46,31 +55,17 @@ namespace nexRemote.Tests
     {
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddDbContextFactory<AppDb>(options =>
-            {
-                options.UseInMemoryDatabase("nexRemote");
-            });
-
-            services.AddScoped(p =>
-                p.GetRequiredService<IDbContextFactory<AppDb>>().CreateDbContext());
+            services.AddDbContext<AppDb, TestingDbContext>();
 
             services.AddIdentity<nexRemoteUser, IdentityRole>(options => options.Stores.MaxLengthForKeys = 128)
              .AddEntityFrameworkStores<AppDb>()
              .AddDefaultUI()
              .AddDefaultTokenProviders();
 
+            services.AddTransient<IAppDbFactory, AppDbFactory>();
             services.AddTransient<IDataService, DataService>();
             services.AddTransient<IApplicationConfig, ApplicationConfig>();
             services.AddTransient<IEmailSenderEx, EmailSenderEx>();
-
-            if (EnvironmentHelper.IsWindows)
-            {
-                services.AddTransient<IDeviceInformationService, DeviceInformationServiceWin>();
-            }
-            else if (EnvironmentHelper.IsLinux)
-            {
-                services.AddTransient<IDeviceInformationService, DeviceInformationServiceLinux>();
-            }
 
             IoCActivator.ServiceProvider = services.BuildServiceProvider();
         }
@@ -79,6 +74,5 @@ namespace nexRemote.Tests
         {
         }
     }
-
 
 }
